# openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/accounts": {
      "get": {
        "operationId": "AccountsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List authorized accounts",
        "tags": ["Accounts"]
      }
    },
    "/api/v1/accounts/{accountId}": {
      "get": {
        "operationId": "AccountsController_getOne",
        "parameters": [
          {
            "name": "accountId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get account details",
        "tags": ["Accounts"]
      }
    },
    "/api/v1/accounts/{accountId}/balances": {
      "get": {
        "operationId": "AccountsController_balances",
        "parameters": [
          {
            "name": "accountId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get account balances",
        "tags": ["Accounts"]
      }
    },
    "/api/v1/accounts/{accountId}/transactions": {
      "get": {
        "operationId": "AccountsController_transactions",
        "parameters": [
          {
            "name": "accountId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Opaque pagination cursor",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 25,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List account transactions",
        "tags": ["Accounts"]
      }
    },
    "/api/v1/accounts/{accountId}/beneficiaries": {
      "get": {
        "operationId": "AccountsController_beneficiaries",
        "parameters": [
          {
            "name": "accountId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List account beneficiaries",
        "tags": ["Accounts"]
      }
    },
...
    "/oauth/authorize": {
      "get": {
        "operationId": "OAuthController_authorize",
        "parameters": [
          {
            "name": "client_id",
            "required": true,
            "in": "query",
            "schema": {
              "example": "client-id",
              "type": "string"
            }
          },
          {
            "name": "redirect_uri",
            "required": true,
            "in": "query",
            "schema": {
              "example": "https://app.example.com/callback",
              "type": "string"
            }
          },
          {
            "name": "response_type",
            "required": true,
            "in": "query",
            "schema": {
              "example": "code",
              "enum": ["code"],
              "type": "string"
            }
          },
          {
            "name": "scope",
            "required": true,
            "in": "query",
            "schema": {
              "example": "accounts:read openid offline_access",
              "type": "string"
            }
          },
          {
            "name": "state",
            "required": true,
            "in": "query",
            "schema": {
              "example": "random-state-value",
              "type": "string"
            }
          },
          {
            "name": "code_challenge",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "code_challenge_method",
            "required": false,
            "in": "query",
            "schema": {
              "enum": ["S256"],
              "type": "string"
            }
          },
          {
            "name": "nonce",
            "required": false,
            "in": "query",
            "description": "OIDC nonce bound into the ID token when openid is requested",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "user_id",
            "required": true,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "institution_id",
            "required": true,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "purpose",
            "required": true,
            "in": "query",
            "schema": {
              "minLength": 3,
              "example": "Account aggregation",
              "type": "string"
            }
          },
          {
            "name": "X-Correlation-Id",
            "in": "header",
            "description": "End-to-end correlation identifier",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "summary": "Create an OAuth 2.1 authorization request (consent bootstrap)",
        "tags": ["OAuth"]
      }
    },
    "/oauth/token": {
      "post": {
        "operationId": "OAuthController_token",
        "parameters": [
          {
            "name": "X-Correlation-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OAuthTokenRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "summary": "Exchange authorization code or refresh token",
        "tags": ["OAuth"]
      }
    },
    "/oauth/revoke": {
      "post": {
        "operationId": "OAuthController_revoke",
        "parameters": [
          {
            "name": "X-Correlation-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OAuthRevokeRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "summary": "Revoke an access or refresh token",
        "tags": ["OAuth"]
      }
    },
    "/oauth/introspect": {
      "post": {
        "operationId": "OAuthController_introspect",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OAuthIntrospectRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "summary": "Introspect token active state and metadata",
        "tags": ["OAuth"]
      }
    },
...
  },
  "info": {
    "title": "Financial API Platform",
    "description": "Secure financial APIs with OAuth 2.1-aligned authorization patterns, OpenID Connect, and consent management",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "OAuth 2.1 access token"
      },
      "idempotency": {
        "type": "apiKey",
        "in": "header",
        "name": "x-idempotency-key",
        "description": "Idempotency key for safe payment retries"
      },
      "correlation": {
        "type": "apiKey",
        "in": "header",
        "name": "x-correlation-id",
        "description": "End-to-end request correlation identifier"
      }
    },
    "schemas": {
...
      "OAuthTokenRequestDto": {
        "type": "object",
        "properties": {
          "grant_type": {
            "type": "string",
            "enum": ["authorization_code", "refresh_token"]
          },
          "client_id": {
            "type": "string",
            "example": "client-id"
          },
          "client_secret": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "redirect_uri": {
            "type": "string"
          },
          "code_verifier": {
            "type": "string"
          },
          "refresh_token": {
            "type": "string"
          }
        },
        "required": ["grant_type", "client_id"]
      },
      "OAuthRevokeRequestDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "token_type_hint": {
            "type": "string",
            "enum": ["access_token", "refresh_token"]
          },
          "client_id": {
            "type": "string"
          }
        },
        "required": ["token", "client_id"]
      },
      "OAuthIntrospectRequestDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "client_id": {
            "type": "string"
          }
        },
        "required": ["token", "client_id"]
      }
    }
  }
}